refactor project
- bump dependencies

- fix circular ref with new keycloak

- fix NPE on thrift claim conversion
