+++ src/main/kotlin/com/rbkmoney/claimmanagementapi/converter/ClaimManagementConverter.kt
@@ -1,17 +1,17 @@
 package com.rbkmoney.claimmanagementapi.converter
 
-import com.rbkmoney.damsel.claim_management.Claim
-import com.rbkmoney.damsel.claim_management.ClaimSearchQuery
-import com.rbkmoney.damsel.claim_management.ClaimStatus
-import com.rbkmoney.damsel.claim_management.ModificationUnit
-import com.rbkmoney.swag.claim_management.model.ClaimChangeset
-import com.rbkmoney.swag.claim_management.model.Modification
+import dev.vality.damsel.claim_management.Claim
+import dev.vality.damsel.claim_management.ClaimSearchQuery
+import dev.vality.damsel.claim_management.ClaimStatus
+import dev.vality.damsel.claim_management.ModificationUnit
+import dev.vality.swag.claim_management.model.ClaimChangeset
+import dev.vality.swag.claim_management.model.Modification
 
 interface ClaimManagementConverter {
-    fun convertClaimToSwag(sourceClaim: Claim): com.rbkmoney.swag.claim_management.model.Claim
-    fun convertClaimListToSwag(sourceClaimList: List<Claim>): List<com.rbkmoney.swag.claim_management.model.Claim>
+    fun convertClaimToSwag(sourceClaim: Claim): dev.vality.swag.claim_management.model.Claim
+    fun convertClaimListToSwag(sourceClaimList: List<Claim>): List<dev.vality.swag.claim_management.model.Claim>
     fun convertClaimChangesetToSwag(changeset: List<ModificationUnit>): ClaimChangeset
-    fun convertModificationUnitToThrift(unitModifications: List<Modification>): List<com.rbkmoney.damsel.claim_management.Modification>
+    fun convertModificationUnitToThrift(unitModifications: List<Modification>): List<dev.vality.damsel.claim_management.Modification>
     fun convertSearchClaimsToThrift(
         partyId: String?,
         claimId: Long?,
