+++ src/main/kotlin/com/rbkmoney/claimmanagementapi/converter/party/contract/ContractAdjustmentModificationUnitConverter.kt
@@ -1,16 +1,16 @@
 package com.rbkmoney.claimmanagementapi.converter.party.contract
 
 import com.rbkmoney.claimmanagementapi.converter.DarkApiConverter
-import com.rbkmoney.damsel.domain.ContractTemplateRef
-import com.rbkmoney.swag.claim_management.model.ContractModification.ContractModificationTypeEnum
+import dev.vality.damsel.domain.ContractTemplateRef
+import dev.vality.swag.claim_management.model.ContractModification.ContractModificationTypeEnum
 import org.springframework.stereotype.Component
-import com.rbkmoney.damsel.claim_management.ContractAdjustmentModification as ThriftContractAdjustmentModification
-import com.rbkmoney.damsel.claim_management.ContractAdjustmentModificationUnit as ThriftContractAdjustmentModificationUnit
-import com.rbkmoney.damsel.claim_management.ContractAdjustmentParams as ThriftContractAdjustmentParams
-import com.rbkmoney.swag.claim_management.model.ContractAdjustmentModification as SwagContractAdjustmentModification
-import com.rbkmoney.swag.claim_management.model.ContractAdjustmentModificationUnit as SwagContractAdjustmentModificationUnit
-import com.rbkmoney.swag.claim_management.model.ContractAdjustmentParams as SwagContractAdjustmentParams
-import com.rbkmoney.swag.claim_management.model.ContractTemplateRef as SwagContractTemplateRef
+import dev.vality.damsel.claim_management.ContractAdjustmentModification as ThriftContractAdjustmentModification
+import dev.vality.damsel.claim_management.ContractAdjustmentModificationUnit as ThriftContractAdjustmentModificationUnit
+import dev.vality.damsel.claim_management.ContractAdjustmentParams as ThriftContractAdjustmentParams
+import dev.vality.swag.claim_management.model.ContractAdjustmentModification as SwagContractAdjustmentModification
+import dev.vality.swag.claim_management.model.ContractAdjustmentModificationUnit as SwagContractAdjustmentModificationUnit
+import dev.vality.swag.claim_management.model.ContractAdjustmentParams as SwagContractAdjustmentParams
+import dev.vality.swag.claim_management.model.ContractTemplateRef as SwagContractTemplateRef
 
 @Component
 class ContractAdjustmentModificationUnitConverter :
