+++ src/main/kotlin/com/rbkmoney/claimmanagementapi/converter/party/contract/ContractModificationUnitConverter.kt
@@ -1,20 +1,20 @@
 package com.rbkmoney.claimmanagementapi.converter.party.contract
 
 import com.rbkmoney.claimmanagementapi.converter.DarkApiConverter
-import com.rbkmoney.damsel.claim_management.ContractTermination
+import dev.vality.damsel.claim_management.ContractTermination
 import org.springframework.stereotype.Component
-import com.rbkmoney.damsel.claim_management.ContractModification as ThriftContractModification
-import com.rbkmoney.damsel.claim_management.ContractModificationUnit as ThriftContractModificationUnit
-import com.rbkmoney.swag.claim_management.model.ContractAdjustmentModificationUnit as SwagContractAdjustmentModificationUnit
-import com.rbkmoney.swag.claim_management.model.ContractContractorModification as SwagContractContractorModification
-import com.rbkmoney.swag.claim_management.model.ContractCreationModification as SwagContractCreationModification
-import com.rbkmoney.swag.claim_management.model.ContractLegalAgreementBindingModification as SwagContractLegalAgreementBindingModification
-import com.rbkmoney.swag.claim_management.model.ContractModification as SwagContractModification
-import com.rbkmoney.swag.claim_management.model.ContractModificationUnit as SwagContractModificationUnit
-import com.rbkmoney.swag.claim_management.model.ContractPayoutToolModificationUnit as SwagContractPayoutToolModificationUnit
-import com.rbkmoney.swag.claim_management.model.ContractReportPreferencesModification as SwagContractReportPreferencesModification
-import com.rbkmoney.swag.claim_management.model.ContractTerminationModification as SwagContractTerminationModification
-import com.rbkmoney.swag.claim_management.model.PartyModificationType as SwagPartyModificationType
+import dev.vality.damsel.claim_management.ContractModification as ThriftContractModification
+import dev.vality.damsel.claim_management.ContractModificationUnit as ThriftContractModificationUnit
+import dev.vality.swag.claim_management.model.ContractAdjustmentModificationUnit as SwagContractAdjustmentModificationUnit
+import dev.vality.swag.claim_management.model.ContractContractorModification as SwagContractContractorModification
+import dev.vality.swag.claim_management.model.ContractCreationModification as SwagContractCreationModification
+import dev.vality.swag.claim_management.model.ContractLegalAgreementBindingModification as SwagContractLegalAgreementBindingModification
+import dev.vality.swag.claim_management.model.ContractModification as SwagContractModification
+import dev.vality.swag.claim_management.model.ContractModificationUnit as SwagContractModificationUnit
+import dev.vality.swag.claim_management.model.ContractPayoutToolModificationUnit as SwagContractPayoutToolModificationUnit
+import dev.vality.swag.claim_management.model.ContractReportPreferencesModification as SwagContractReportPreferencesModification
+import dev.vality.swag.claim_management.model.ContractTerminationModification as SwagContractTerminationModification
+import dev.vality.swag.claim_management.model.PartyModificationType as SwagPartyModificationType
 
 @Component
 class ContractModificationUnitConverter(
