+++ src/main/kotlin/com/rbkmoney/claimmanagementapi/converter/party/contract/PayoutToolInfoConverter.kt
@@ -1,16 +1,16 @@
 package com.rbkmoney.claimmanagementapi.converter.party.contract
 
 import com.rbkmoney.claimmanagementapi.converter.DarkApiConverter
-import com.rbkmoney.damsel.domain.InternationalBankAccount
-import com.rbkmoney.damsel.domain.PaymentInstitutionAccount
+import dev.vality.damsel.domain.InternationalBankAccount
+import dev.vality.damsel.domain.PaymentInstitutionAccount
 import org.springframework.stereotype.Component
-import com.rbkmoney.damsel.domain.PayoutToolInfo as ThriftPayoutToolInfo
-import com.rbkmoney.damsel.domain.WalletInfo as ThriftWalletInfo
-import com.rbkmoney.swag.claim_management.model.InternationalBankAccount as SwagInternationalBankAccount
-import com.rbkmoney.swag.claim_management.model.PaymentInstitutionAccount as SwagPaymentInstitutionAccount
-import com.rbkmoney.swag.claim_management.model.PayoutToolInfo as SwagPayoutToolInfo
-import com.rbkmoney.swag.claim_management.model.RussianBankAccount as SwagRussianBankAccount
-import com.rbkmoney.swag.claim_management.model.WalletInfo as SwagWalletInfo
+import dev.vality.damsel.domain.PayoutToolInfo as ThriftPayoutToolInfo
+import dev.vality.damsel.domain.WalletInfo as ThriftWalletInfo
+import dev.vality.swag.claim_management.model.InternationalBankAccount as SwagInternationalBankAccount
+import dev.vality.swag.claim_management.model.PaymentInstitutionAccount as SwagPaymentInstitutionAccount
+import dev.vality.swag.claim_management.model.PayoutToolInfo as SwagPayoutToolInfo
+import dev.vality.swag.claim_management.model.RussianBankAccount as SwagRussianBankAccount
+import dev.vality.swag.claim_management.model.WalletInfo as SwagWalletInfo
 
 @Component
 class PayoutToolInfoConverter(
