+++ src/main/kotlin/com/rbkmoney/claimmanagementapi/converter/party/contract/PayoutToolModificationUnitConverter.kt
@@ -1,18 +1,18 @@
 package com.rbkmoney.claimmanagementapi.converter.party.contract
 
 import com.rbkmoney.claimmanagementapi.converter.DarkApiConverter
-import com.rbkmoney.damsel.claim_management.PayoutToolModification
-import com.rbkmoney.damsel.claim_management.PayoutToolParams
-import com.rbkmoney.damsel.domain.CurrencyRef
-import com.rbkmoney.swag.claim_management.model.ContractPayoutToolModification.PayoutToolModificationTypeEnum.CONTRACTPAYOUTTOOLCREATIONMODIFICATION
-import com.rbkmoney.swag.claim_management.model.ContractPayoutToolModification.PayoutToolModificationTypeEnum.CONTRACTPAYOUTTOOLINFOMODIFICATION
+import dev.vality.damsel.claim_management.PayoutToolModification
+import dev.vality.damsel.claim_management.PayoutToolParams
+import dev.vality.damsel.domain.CurrencyRef
+import dev.vality.swag.claim_management.model.ContractPayoutToolModification.PayoutToolModificationTypeEnum.CONTRACTPAYOUTTOOLCREATIONMODIFICATION
+import dev.vality.swag.claim_management.model.ContractPayoutToolModification.PayoutToolModificationTypeEnum.CONTRACTPAYOUTTOOLINFOMODIFICATION
 import org.springframework.stereotype.Component
-import com.rbkmoney.damsel.claim_management.PayoutToolModificationUnit as ThriftPayoutToolModificationUnit
-import com.rbkmoney.swag.claim_management.model.ContractModification as SwagContractModification
-import com.rbkmoney.swag.claim_management.model.ContractPayoutToolCreationModification as SwagContractPayoutToolCreationModification
-import com.rbkmoney.swag.claim_management.model.ContractPayoutToolInfoModification as SwagContractPayoutToolInfoModification
-import com.rbkmoney.swag.claim_management.model.ContractPayoutToolModificationUnit as SwagContractPayoutToolModificationUnit
-import com.rbkmoney.swag.claim_management.model.CurrencyRef as SwagCurrencyRef
+import dev.vality.damsel.claim_management.PayoutToolModificationUnit as ThriftPayoutToolModificationUnit
+import dev.vality.swag.claim_management.model.ContractModification as SwagContractModification
+import dev.vality.swag.claim_management.model.ContractPayoutToolCreationModification as SwagContractPayoutToolCreationModification
+import dev.vality.swag.claim_management.model.ContractPayoutToolInfoModification as SwagContractPayoutToolInfoModification
+import dev.vality.swag.claim_management.model.ContractPayoutToolModificationUnit as SwagContractPayoutToolModificationUnit
+import dev.vality.swag.claim_management.model.CurrencyRef as SwagCurrencyRef
 
 @Component
 class PayoutToolModificationUnitConverter(
