+++ src/main/kotlin/com/rbkmoney/claimmanagementapi/converter/party/contractor/ContractorIdentificationLevelConverter.kt
@@ -1,10 +1,10 @@
 package com.rbkmoney.claimmanagementapi.converter.party.contractor
 
 import com.rbkmoney.claimmanagementapi.converter.DarkApiConverter
-import com.rbkmoney.swag.claim_management.model.ContractorModification
+import dev.vality.swag.claim_management.model.ContractorModification
 import org.springframework.stereotype.Component
-import com.rbkmoney.damsel.domain.ContractorIdentificationLevel as ThriftContractorIdentificationLevel
-import com.rbkmoney.swag.claim_management.model.ContractorIdentificationLevel as SwagContractorIdentificationLevel
+import dev.vality.damsel.domain.ContractorIdentificationLevel as ThriftContractorIdentificationLevel
+import dev.vality.swag.claim_management.model.ContractorIdentificationLevel as SwagContractorIdentificationLevel
 
 @Component
 class ContractorIdentificationLevelConverter :
