+++ src/main/kotlin/com/rbkmoney/claimmanagementapi/service/PartyManagementService.kt
@@ -3,10 +3,10 @@ package com.rbkmoney.claimmanagementapi.service
 import com.rbkmoney.claimmanagementapi.exception.client.ForbiddenException
 import com.rbkmoney.claimmanagementapi.exception.server.DarkApi5xxException
 import com.rbkmoney.claimmanagementapi.security.KeycloakService
-import com.rbkmoney.damsel.payment_processing.InternalUser
-import com.rbkmoney.damsel.payment_processing.PartyManagementSrv
-import com.rbkmoney.damsel.payment_processing.UserInfo
-import com.rbkmoney.damsel.payment_processing.UserType
+import dev.vality.damsel.payment_processing.InternalUser
+import dev.vality.damsel.payment_processing.PartyManagementSrv
+import dev.vality.damsel.payment_processing.UserInfo
+import dev.vality.damsel.payment_processing.UserType
 import mu.KotlinLogging
 import org.springframework.stereotype.Service
